% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{augment.rehydra_analysis}
\alias{augment.rehydra_analysis}
\title{Augment Method for rehydra Analysis Objects}
\usage{
\method{augment}{rehydra_analysis}(x, ...)
}
\arguments{
\item{x}{A \code{rehydra_analysis} object.}

\item{...}{Not used.}
}
\value{
A tibble joining segmented trajectories and per-cycle indices.
}
\description{
Augment Method for rehydra Analysis Objects
}
