% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damage_functions.R
\name{damage_metrics}
\alias{damage_metrics}
\title{Calculate Drought Damage Metrics}
\usage{
damage_metrics(
  data,
  time = time,
  value = transformed_value,
  group_by = c("genotype", "treatment", "replicate", "plant_id", "variable", "cycle"),
  threshold = NULL
)
}
\arguments{
\item{data}{Segmented data from \code{\link[=segment_drought_cycle]{segment_drought_cycle()}}.}

\item{time}{Time column.}

\item{value}{Value column used for calculations.}

\item{group_by}{Grouping columns.}

\item{threshold}{Relative drop threshold used to compute
\code{time_to_damage}. If \code{NULL}, time to the absolute minimum is used.}
}
\value{
A tibble of drought-damage metrics by group and cycle.
}
\description{
Quantify intensity and dynamics of physiological damage during drought.
}
\details{
Metrics include:
\itemize{
\item \code{drought_minimum}
\item \code{drought_drop_absolute = PreDrought - DroughtMinimum}
\item \code{drought_drop_relative = (PreDrought - DroughtMinimum) / PreDrought}
\item \code{damage_rate = (DroughtMinimum - PreDrought) / duration_drought}
\item \code{max_decline_rate} (minimum first derivative during drought)
\item \verb{damage_auc = integral(PreDrought_reference - observed_value) dt}
\item \verb{normalized_damage_auc = damage_auc / (duration_drought * |PreDrought|)}
\item \code{time_to_damage}
}
}
\examples{
data(rehydra_data)
prepared <- prepare_rehydra_data(
  data = rehydra_data,
  time = time_days,
  genotype = genotype,
  treatment = treatment,
  replicate = replicate,
  variables = c(water_potential, stomatal_conductance)
)
segments <- segment_drought_cycle(prepared, time = time, value = transformed_value)
damage_metrics(segments)
}
\references{
Lloret et al. (2011) \url{doi:10.1111/j.1600-0706.2011.19372.x}.
}
