% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_trend.R
\name{memory_direction}
\alias{memory_direction}
\title{Direction of Improvement for an Index}
\usage{
memory_direction(metric)
}
\arguments{
\item{metric}{Character vector of metric names.}
}
\value{
A logical vector, \code{TRUE} where a lower value is better.
}
\description{
Report whether a lower value of a given rehydra index means a better
physiological outcome. Used by \code{\link[=memory_trend]{memory_trend()}} to sign its memory classes.
}
\examples{
memory_direction(c("Rs", "deficit_auc", "t50"))
}
