% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{plot.rehydra_analysis}
\alias{plot.rehydra_analysis}
\title{Plot Method for rehydra Analysis Objects}
\usage{
\method{plot}{rehydra_analysis}(x, ...)
}
\arguments{
\item{x}{A \code{rehydra_analysis} object.}

\item{...}{Additional arguments forwarded to
\code{\link[=plot_recovery_trajectory]{plot_recovery_trajectory()}}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Method for rehydra Analysis Objects
}
