% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_damage_recovery}
\alias{plot_damage_recovery}
\title{Plot Damage, Recovery, and Residual Costs}
\usage{
plot_damage_recovery(
  damage_results,
  recovery_results,
  facet_by = c("genotype", "variable")
)
}
\arguments{
\item{damage_results}{Output from \code{\link[=damage_metrics]{damage_metrics()}}.}

\item{recovery_results}{Output from \code{\link[=recovery_metrics]{recovery_metrics()}}.}

\item{facet_by}{Faceting columns.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Damage, Recovery, and Residual Costs
}
