% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_memory_effect}
\alias{plot_memory_effect}
\title{Plot Memory Effect Metrics}
\usage{
plot_memory_effect(memory_results, facet_by = c("treatment", "variable"))
}
\arguments{
\item{memory_results}{Output from \code{\link[=memory_effect]{memory_effect()}}.}

\item{facet_by}{Faceting columns.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Memory Effect Metrics
}
