% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_resilience_indices}
\alias{plot_resilience_indices}
\title{Plot Resilience Indices}
\usage{
plot_resilience_indices(
  resilience_results,
  facet_by = c("treatment", "variable")
)
}
\arguments{
\item{resilience_results}{Output from \code{\link[=resilience_index]{resilience_index()}}.}

\item{facet_by}{Faceting columns.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Resilience Indices
}
