% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_stability}
\alias{plot_stability}
\title{Plot Stability Metrics}
\usage{
plot_stability(stability_results, facet_by = c("treatment", "variable"))
}
\arguments{
\item{stability_results}{Output from \code{\link[=stability_index]{stability_index()}}.}

\item{facet_by}{Faceting columns.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Stability Metrics
}
