% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_functions.R
\name{print.rehydra_anova}
\alias{print.rehydra_anova}
\title{Print Method for rehydra ANOVA Objects}
\usage{
\method{print}{rehydra_anova}(x, ...)
}
\arguments{
\item{x}{A \code{rehydra_anova} object.}

\item{...}{Not used.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for rehydra ANOVA Objects
}
