% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_rehydra_app}
\alias{run_rehydra_app}
\title{Run the rehydra Shiny Application}
\usage{
run_rehydra_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
The Shiny app object invisibly after launch.
}
\description{
Launch the interactive Shiny interface bundled with \code{rehydra}.
}
\examples{
\dontrun{
run_rehydra_app()
}
}
