% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{summary.rehydra_analysis}
\alias{summary.rehydra_analysis}
\title{Summary Method for rehydra Analysis Objects}
\usage{
\method{summary}{rehydra_analysis}(object, ...)
}
\arguments{
\item{object}{A \code{rehydra_analysis} object.}

\item{...}{Not used.}
}
\value{
A one-row tibble with object dimensions.
}
\description{
Summary Method for rehydra Analysis Objects
}
