% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_functions.R
\name{summary.rehydra_anova}
\alias{summary.rehydra_anova}
\title{Summary Method for rehydra ANOVA Objects}
\usage{
\method{summary}{rehydra_anova}(object, ...)
}
\arguments{
\item{object}{A \code{rehydra_anova} object.}

\item{...}{Not used.}
}
\value{
The ANOVA table as a tibble.
}
\description{
Summary Method for rehydra ANOVA Objects
}
