% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{tidy.rehydra_analysis}
\alias{tidy.rehydra_analysis}
\title{Tidy Method for rehydra Analysis Objects}
\usage{
\method{tidy}{rehydra_analysis}(x, ...)
}
\arguments{
\item{x}{A \code{rehydra_analysis} object.}

\item{...}{Not used.}
}
\value{
A tibble combining resilience and memory metrics.
}
\description{
Tidy Method for rehydra Analysis Objects
}
