% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_functions.R
\name{tidy.rehydra_anova}
\alias{tidy.rehydra_anova}
\title{Tidy Method for rehydra ANOVA Objects}
\usage{
\method{tidy}{rehydra_anova}(x, ...)
}
\arguments{
\item{x}{A \code{rehydra_anova} object.}

\item{...}{Not used.}
}
\value{
The ANOVA table as a tibble.
}
\description{
Tidy Method for rehydra ANOVA Objects
}
